backend/requirements_django.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-decouple==3.8
django-extensions==3.2.3
django-filter==24.2
djangorestframework-simplejwt==5.3.0
Pillow==10.3.0
shortuuid==1.0.13
google-generativeai==0.5.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.3.6
django-redis==5.4.0
gunicorn==22.0.0
whitenoise==6.6.0
